refactor: simplify v5 to v6 transfer
