+++ cyberdrop_dl/managers/manager.py
@@ -22,7 +22,7 @@ from cyberdrop_dl.utils import constants
 from cyberdrop_dl.utils.args import ParsedArgs
 from cyberdrop_dl.utils.data_enums_classes.supported_domains import SUPPORTED_FORUMS
 from cyberdrop_dl.utils.logger import log
-from cyberdrop_dl.utils.transfer.db_setup import TransitionManager
+from cyberdrop_dl.utils.transfer import transfer_v5_db_to_v6
 
 if TYPE_CHECKING:
     from asyncio import TaskGroup
@@ -46,8 +46,6 @@ class Manager:
         self.download_manager: DownloadManager = field(init=False)
         self.progress_manager: ProgressManager = field(init=False)
         self.live_manager: LiveManager = field(init=False)
-
-        self.first_time_setup: TransitionManager = TransitionManager(self)
 
         self._loaded_args_config: bool = False
         self._made_portable: bool = False
@@ -123,7 +121,7 @@ class Manager:
         if not isinstance(self.db_manager, DBManager):
             self.db_manager = DBManager(self, self.path_manager.history_db)
             await self.db_manager.startup()
-        self.first_time_setup.transfer_v5_to_new_hashtable()
+        transfer_v5_db_to_v6(self.path_manager.history_db)
         if not isinstance(self.hash_manager, HashManager):
             self.hash_manager = HashManager(self)
             await self.hash_manager.startup()
